requirements.txt
PyYAML==6.0
requests==2.28.2
click==8.1.3